feat:opacity in sections
