--- src/functions/createBooking.js
@@ -1,15 +1,0 @@
-
-
-function createBooking(customerObj, userDate, roomNumber){
-  if(!userDate){return 'please enter a date.'} 
-  if(!roomNumber){return 'please choose a room.'} 
-    let booking = {
-      userID: customerObj.id, 
-      date: userDate, 
-      roomNumber: roomNumber
-    }
-    return booking
-}
-
-
-export { createBooking }
